Just starting to get the IMU data
Right now, I'm grabbing the data and publishing it, but it doesn't seem right at all.
Do I need a calibration of some kind? Are the axes/order of the fields in the default
poser process wrong? Maybe Adam knows something about this? Or maybe there is something
on discord. Definitely not right as it is.