Quick fix to try and address issue #2
Issue OP mentioned is almost certainly due to changes from OpenCV 2.x to OpenCV
3.x. It seems that in the switch from 2.x to 3.x they re-arranged the headers
and also made it such that the headers are not able to be compiled by a C
compiler. One option would be to replace all C++ API's used in this project with
C APIs, but it seemed easier to just break apart the headers required and
compile the C parts and the C++ parts as separate libraries.

Read more:

	+ An example of the OpenCV change can be seen here in the core.hpp
	  file on tag 3.4:
	   https://github.com/opencv/opencv/blob/3.4.0/modules/core/include/opencv2/core.hpp#L48-L50
	    Compare that to the corresponding file on tag 2.4.8
		 https://github.com/opencv/opencv/blob/2.4.8.1/modules/core/include/opencv2/core/core.hpp
	+ Top of the 2-to-3 migration guide https://docs.opencv.org/master/db/dfa/tutorial_transition_guide.html